requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
google-generativeai==0.8.3
pydantic==2.5.0
python-dotenv==1.0.0